// src/app/enums/store.ts
export enum ActionTypes {
  GET_TWIN_STORE = "[LDIGITAL] Get",
  UPDATE_TWIN_STORE = "[LDIGITAL] Update",
  ADD_NOTIFICATION = "[LDIGITAL] notification",
  ASSIGN_ELEMENT = "[LDIGITAL] assignelement",
  TASKS = "[LDIGITAL] tasks",
  SEARCH_DATA = "[LDIGITAL] search",
  ASSIGN_TASKS = "[LDIGITAL] assigntasks",
  TOTAL_PROJECTS = "[LDIGITAL] totalprojects",
  PROJECT_CREATE = "[LDIGITAL] project create",
  PROJECT_STEP = "[LDIGITAL] project step",
  CREATE_META_DATA = "[LDIGITAL] createn meta data step",
  PROJECT_ID_CREATION = "[LDIGITAL] creating project ID",
  ERROR_LOAD = "[LDIGITAL] creating project ID",
  CURRENT_STEP = "[LDIGITAL] tracking step",
  TEMPLATE = "[LDIGITAL] handling template",
  TAST_ASSIGNMENT = "[LDIGITAL] taskSelectedAsssignedmentElements ",
  JOB_LIST = "[LDIGITAL] Job list template",
  CHECK_CONFIGURATION = "[LDIGITAL] configuration tests",
  CISV = "[LDIGITAL] CISV list template",

  FLOWDATA = "[LDIGITAL] FLOWDATA list template",
  SCENATIO = "[LDIGITAL SCENARIO]",
  PAGESCENATIO = "[LDIGIGTAL PAGESCENARIO]",
  UPDATEPROGRAM = "[LDigital UPDATEPROGRAM]",
}